스토어 평균 별점 좋아요 수 패치 완료

+++ src/main/java/kr/co/hanipactor/application/store/model/StoreGetListRes.java
@@ -19,7 +19,7 @@ public class StoreGetListRes {
     private int maxDeliveryFee; // 최대 배달 요금
     private int minDeliveryFee; // 최소 배달 요금
     private int minAmount; // 최소 주문 금액
-    private int rating; // 별점 평균
+    private double rating; // 별점 평균
     private int favorites; // 좋아요 갯수
     private List<EnumStoreCategory> category; // 카테고리
 }
